feat: HostBridge trait impl
